long -> int

+++ tabdiff/models/noise_schedule.py
@@ -194,7 +194,7 @@ class TreeLayeredNoiseNum(nn.Module):
     # NEW CODE:
     # d_t: the layer that we diffusing
     # d_t = ((1-t) * self.num_tree_layers).int()
-    d_t = torch.floor((1 - t) * self.num_tree_layers).long()
+    d_t = torch.floor((1 - t) * self.num_tree_layers).int()
     # d_t = d_t.clamp(min=0, max=self.num_tree_layers - 1)
 
     # t_ is the shifted and normalized noise t for the layer d_t
